Move taxe_caluclation into a seperate service
